restructure general.js
