+++ functions/copySign.js
@@ -1,0 +1,7 @@
+var sign1 = require('./sign1');
+
+module.exports = function copySign (value, sign) {
+    "use strict";
+
+    return Math.abs(value) * sign1(sign);
+};
