+++ functions/roundToPrecision.js
@@ -1,0 +1,7 @@
+module.exports = function roundToPrecision (value, precision) {
+    "use strict";
+
+    var factor = Math.pow(10, ~~precision);
+
+    return Math.round(value * factor) / factor;
+};
